requirements.txt
absl-py==0.7.1
albumentations==0.5.2
appnope==0.1.0
astor==0.7.1
astroid==2.4.2
attrs==20.2.0
backcall==0.1.0
bleach==3.1.0
certifi==2020.6.20
chardet==3.0.4
confuse==1.3.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.0
defusedxml==0.5.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
future==0.17.1
gast==0.2.2
gmplot==1.2.0
google-streetview==1.2.9
grpcio==1.19.0
h5py==2.9.0
htmlmin==0.1.12
idna==2.8
ImageHash==4.1.0
imageio==2.5.0
imgaug==0.4.0
imgkit==1.0.2
imutils==0.5.2
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.13.3
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
kwconfig==1.1.7
lazy-object-proxy==1.4.3
llvmlite==0.34.0
lxml==4.6.2
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mock==2.0.0
munch==2.5.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.5
notebook==5.7.6
numba==0.51.2
numpy==1.16.2
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.1.3
pandas-profiling==2.9.0
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pexpect==4.6.0
phik==0.10.0
pickleshare==0.7.5
Pillow==8.0.1
pretrainedmodels==0.7.4
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.5.3
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.24.0
scikit-image==0.17.2
scikit-learn==0.20.3
scipy==1.5.3
seaborn==0.11.0
segmentation-models-pytorch==0.1.3
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.1
SimpleITK==2.0.2
six==1.12.0
sklearn==0.0
tangled-up-in-unicode==0.0.6
tb-nightly==1.14.0a20190528
tensorboard==2.0.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tifffile==2020.9.3
timm==0.3.2
toml==0.10.1
torch==1.7.1
torchvision==0.4.1
tornado==6.0.2
tqdm==4.50.2
traitlets==4.3.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.24.1
visions==0.5.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.5.1
wrapt==1.12.1
